# hotel-order-ui/src/app/services/models/item-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ItemRequest {
  categoryId: string;
  description: string;
  id?: string;
  imageUrl?: string;
  isAvailable: boolean;
  name: string;
  popularityIndex?: number;
  price: number;
  stockQuantity?: number;
}